(refactor) moved decease panel to patientChart
