+++ packages/esm-morgue-app/src/card/compartment.scss
@@ -88,3 +88,9 @@
   width: 100%;
   margin-top: layout.$spacing-05;
 }
+
+.configurableLink {
+  margin: 0;
+  padding: 0;
+  text-decoration: none;
+}
