+++ packages/esm-morgue-app/src/config-schema.ts
@@ -1,23 +1,5 @@
 import { Type } from '@openmrs/esm-framework';
 
-export const configSchema = {
-  autopsyEncounterUuid: {
-    _type: Type.String,
-    _description: 'Encounter UUID for autopsy',
-    _default: '',
-  },
-  formsList: {
-    _type: Type.Object,
-    _description: 'List of form UUIDs',
-    _default: {
-      autopsyFormUuid: '',
-    },
-  },
-};
+export const configSchema = {};
 
-export type ConfigObject = {
-  formsList: {
-    autopsyFormUuid: string;
-  };
-  autopsyEncounterUuid: string;
-};
+export type ConfigObject = {};
