+++ packages/esm-morgue-app/src/root.component.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import MainComponent from './component/main.component';
-import DeceasedDetailsView from './details/deceased-details.component';
 import { AdmittedQueue } from './tables/admitted-queue.component';
 
 const Root: React.FC = () => {
@@ -12,7 +11,6 @@ const Root: React.FC = () => {
       <Routes>
         <Route path="/" element={<MainComponent />} />
         <Route path="/allocation" element={<AdmittedQueue />} />
-        <Route path="/patient/:patientUuid" element={<DeceasedDetailsView />} />
       </Routes>
     </BrowserRouter>
   );
