+++ packages/esm-morgue-app/src/tabs/tabs.scss
@@ -12,8 +12,6 @@
 }
 
 .serviceContainer {
-  // background-color: $ui-02;
-  // border: 1px solid $ui-03;
   width: 100%;
   margin: 0 auto;
   max-width: 95vw;
@@ -129,7 +127,6 @@
 
 .content {
   @include type.type-style('heading-compact-01');
-  // color: $text-02;
   margin-top: layout.$spacing-05;
   margin-bottom: layout.$spacing-03;
 }
@@ -142,7 +139,6 @@
 
   h4 {
     @include type.type-style('heading-compact-02');
-    // color: $text-02;
 
     &:after {
       content: '';
@@ -150,14 +146,12 @@
       width: 2rem;
       padding-top: 3px;
       border-bottom: 0.375rem solid;
-      // @include brand-03(border-bottom-color);
     }
   }
 }
 
 .tile {
   text-align: center;
-  // border: 1px solid $ui-03;
 }
 
 .menuitem {
@@ -179,13 +173,11 @@
 
 .filterEmptyStateContent {
   @include type.type-style('heading-compact-02');
-  // color: $text-02;
   margin-bottom: 0.5rem;
 }
 
 .filterEmptyStateHelper {
   @include type.type-style('body-compact-01');
-  // color: $text-02;
 }
 
 .actionBtn {
