+++ packages/esm-morgue-app/src/types/index.ts
@@ -34,32 +34,6 @@ export interface DeceasedPatientResponse {
   results: Patient[];
 }
 
-export interface Observation {
-  uuid: string;
-  concept: {
-    uuid: string;
-    display: string;
-    conceptClass: {
-      uuid: string;
-      display: string;
-    };
-  };
-  display: string;
-  groupMembers: null | Array<{
-    uuid: string;
-    concept: {
-      uuid: string;
-      display: string;
-    };
-    value: {
-      uuid: string;
-      display: string;
-    };
-    display: string;
-  }>;
-  value: any;
-  obsDatetime?: string;
-}
 export interface DeceasedInfo {
   uuid: string;
   display: string;
