--- packages/esm-morgue-app/src/utils/utils.ts
@@ -1,5 +1,0 @@
-import { makeUrl } from '@openmrs/esm-framework';
-
-export function makeUrlUrl(path: string) {
-  return new URL(makeUrl(path), window.location.toString());
-}
