--- packages/esm-morgue-app/src/workspaces/formEntry.workspace.tsx
@@ -1,57 +1,0 @@
-import { DefaultWorkspaceProps, ExtensionSlot, useConnectivity, usePatient } from '@openmrs/esm-framework';
-import React, { useMemo } from 'react';
-
-type FormEntryWorkspaceProps = DefaultWorkspaceProps & {
-  formUuid?: string;
-  patientUuid?: string;
-  encounterUuid?: string;
-  mutateForm: () => void;
-};
-
-const FormEntryWorkspace: React.FC<FormEntryWorkspaceProps> = (props) => {
-  const { formUuid, patientUuid, encounterUuid, mutateForm, closeWorkspace, closeWorkspaceWithSavedChanges } = props;
-  const { patient, isLoading } = usePatient(patientUuid);
-  const isOnline = useConnectivity();
-  const state = useMemo(
-    () => ({
-      ...props,
-      view: 'form',
-      formUuid: formUuid ?? null,
-      visitUuid: '',
-      visitTypeUuid: '',
-      visitStartDatetime: null,
-      visitStopDatetime: null,
-      isOffline: !isOnline,
-      patientUuid: patientUuid ?? null,
-      patient,
-      encounterUuid: encounterUuid ?? null,
-      closeWorkspace: () => {
-        typeof mutateForm === 'function' && mutateForm();
-        closeWorkspace();
-      },
-      closeWorkspaceWithSavedChanges: () => {
-        typeof mutateForm === 'function' && mutateForm();
-        closeWorkspaceWithSavedChanges();
-      },
-    }),
-    [
-      props,
-      formUuid,
-      isOnline,
-      patientUuid,
-      patient,
-      encounterUuid,
-      mutateForm,
-      closeWorkspace,
-      closeWorkspaceWithSavedChanges,
-    ],
-  );
-
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
-
-  return <ExtensionSlot name="form-widget-slot" state={state} />;
-};
-
-export default FormEntryWorkspace;
