+++ packages/esm-patient-clinical-view-app/src/deceased-panel/panels/attachement.component.tsx
