+++ packages/esm-patient-clinical-view-app/src/deceased-panel/panels/panels.scss
