+++ packages/esm-patient-clinical-view-app/src/deceased-panel/tabs/tabs.scss
