+++ packages/esm-patient-clinical-view-app/src/index.ts
@@ -13,6 +13,7 @@ import {
   caseManagementDashboardMeta,
   contactListDashboardMeta,
   familyHistoryDashboardMeta,
+  morgueDeceasedDetailsDashboardMeta,
   otherRelationshipsDashboardMeta,
   peerCalendarDashboardMeta,
   relationshipsDashboardMeta,
@@ -52,6 +53,7 @@ import PeerForm from './peer-calendar/forms/peer-form.workspace';
 import FormEntryWorkspace from './peer-calendar/forms/form-entry.workspace';
 import RelationshipUpdateForm from './relationships/forms/relationships-update-form.workspace';
 import DeleteRelationshipConfirmDialog from './relationships/modals/delete-relationship-dialog.modal';
+import DeceasedDetailsView from './deceased-panel/tabs/tabs.component';
 
 const moduleName = '@kenyaemr/esm-patient-clinical-view-app';
 
@@ -132,6 +134,14 @@ export const caseManagementForm = getSyncLifecycle(CaseManagementForm, options);
 export const caseEncounterDashboardLink = getSyncLifecycle(createDashboardLink(caseEncounterDashboardMeta), options);
 // export const caseEncounterHeader = getSyncLifecycle(CaseEncounterHeader, options);
 export const caseEncounterTable = getSyncLifecycle(CaseEncounterOverviewComponent, options);
+
+// Morgue deceased details
+export const deceasedPanelDashboardLink = getSyncLifecycle(
+  createDashboardLink(morgueDeceasedDetailsDashboardMeta),
+  options,
+);
+
+export const deceasedDetailsTabs = getSyncLifecycle(DeceasedDetailsView, options);
 
 // Dashboard links for Maternal and Child Health services
 export const antenatalCareLink = getSyncLifecycle(createDashboardLink(antenatalDashboardMeta), options);
